Move ChatHistoryUtils to ChatRepository
- Adds extractSharedViewInIds to ChatRepository interface.
- Implements logic in ChatRepositoryImpl.
- Updates ChatViewModel to pass sharedViewInIds via ChatHistoryScreenData.
- Refactors ChatHistoryFragment and ChatHistoryAdapter to use passed map.
- Moves corresponding tests to ChatRepositoryImplTest.
- Deletes ChatHistoryUtils.kt and ChatHistoryUtilsTest.kt.
